Model: remove LSMM2 and add it to .gitignore
Most users will need to recompile the LSMM2 binary to make it
work with their local machine. Let's remove the binary from version
control so that users do not try to use the compiled binary from
@kallisons' machine.

Also add it to .gitignore, so that the compiled binary is not
suggested as a change.